Extract useProjects hook
